Fixed bad parameter values and more additions
finalTime and keepEvery parameters in the sampling function of the run_case_neutral.R script were set in fixed values instead of passing the parameters declared at the beginning ofthe script. A column for the number of drivers mutations detected is included in the sampling output. Output as a tibble.