Change project structure to fit an R package
The project is going to be build as an R package to allow further reproducibility and better documentation. Using devtools we created the structure of a package, renaming pNpS_sims to pNpS.sims all the project. I added a DESCRIPTION file. NSsampling functions was moved to the new R/ directory and renamed to extract_samples. Added also new roxygen documentation to the function and the package.
Closes #12