Improve function for sampling
Improve the extract_sampling function to allow different requested
features. Pass the options stored from the simulation as a list in order
to allow automatization. Add more parameters to the output tibble like
number of deletereous mutations or mean fitness in the population. Run
a bunch of different frequency thresholds to avoid wasting time with
tcrossprod. Reorganize the output. Add documentation.

Resolves: #8 #11 #13